+++ R/ClusterR.R
@@ -2,4 +2,5 @@
 #' @importFrom Rcpp evalCpp
 #' @import gtools
 #' @importFrom utils globalVariables
+#' @importFrom lifecycle badge deprecate_warn
 NULL