+++ R/clustering_functions.R
@@ -1089,7 +1089,7 @@ predict_MBatchKMeans = function(data, CENTROIDS, fuzzy = FALSE) {
 #' @param swap_phase either TRUE or FALSE. If TRUE then both phases ('build' and 'swap') will take place. The 'swap_phase' is considered more computationally intensive.
 #' @param fuzzy either TRUE or FALSE. If TRUE, then probabilities for each cluster will be returned based on the distance between observations and medoids
 #' @param verbose either TRUE or FALSE, indicating whether progress is printed during clustering
-#' @param seed integer value for random number generator (RNG)
+#' @param seed `r lifecycle::badge("deprecated")` `seed` (integer value for random number generator (RNG)) is no longer supported and will be removed in version 1.3.0
 #' @return a list with the following attributes: medoids, medoid_indices, best_dissimilarity, dissimilarity_matrix, clusters, fuzzy_probs (if fuzzy = TRUE), silhouette_matrix, clustering_stats
 #' @author Lampros Mouselimis
 #' @details
@@ -1109,6 +1109,14 @@ predict_MBatchKMeans = function(data, CENTROIDS, fuzzy = FALSE) {
 #' cm = Cluster_Medoids(dat, clusters = 3, distance_metric = 'euclidean', swap_phase = TRUE)
 #'
 Cluster_Medoids = function(data, clusters, distance_metric = 'euclidean', minkowski_p = 1.0, threads = 1, swap_phase = TRUE, fuzzy = FALSE, verbose = FALSE, seed = 1) {
+
+  if (seed != 1) {
+    lifecycle::deprecate_warn(
+      when = "1.2.6",
+      what = "Cluster_Medoids(seed)",
+      details = "Ability to set a seed value will be dropped in version 1.3.0"
+    )
+  }
 
   if ('data.frame' %in% class(data)) data = as.matrix(data)
   if (!inherits(data, 'matrix')) stop('data should be either a matrix or a data frame or a dissimilarity matrix with equal number of rows and columns and a diagonal equal to 0.0')